fix typing
